wip: scraping all, removing gets in project requests
