+++ fiis/scrapes/investidor10.py
@@ -9,44 +9,15 @@ import pandas as pd
 from bs4 import BeautifulSoup
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from fiis.constants import INVESTIDOR10_BASE_URL, HEADERS, FILE_NAME
+from fiis.constants import INVESTIDOR10_BASE_URL, HEADERS
+from fiis.utils import write_csv_file
 
 
 log_format = '%(asctime)s - %(levelname)s - %(message)s'
 logging.basicConfig(format=log_format, level=logging.INFO)
 
 
-def get_downloads_path() -> str:
-    '''Retorna o caminho da pasta downloads, um nível acima do diretório do script.'''
-    script_dir = os.path.dirname(os.path.abspath(__file__))
-    parent_dir = os.path.dirname(script_dir)
-    return os.path.join(parent_dir, 'downloads')
-
-
-def ensure_downloads_folder() -> None:
-    '''Cria a pasta 'downloads' um nível acima do diretório do script, se não existir.'''
-    downloads_path = get_downloads_path()
-
-    if not os.path.exists(downloads_path):
-        os.makedirs(downloads_path)
-        logging.info(f'Pasta "downloads" criada em: {downloads_path}')
-
-
-def write_csv_file(data: pd.DataFrame, file_name: str) -> None:
-    ''' 
-    Escreve um arquivo CSV com os dados fornecidos, escrita feita na pasta downloads.
-
-    Args:
-        data (pd.DataFrame): DataFrame contendo os dados.
-        file_name (str): Nome do arquivo CSV.
-    '''
-    ensure_downloads_folder()
-    downloads_path = get_downloads_path()
-    data.to_csv(f'{downloads_path}/{file_name}.csv', index=False)
-    logging.info(f'Arquivo {file_name}.csv escrito com sucesso!')
-
-
-class FIIsScraper:
+class Investidor10Scraper:
     ''' Classe para obter dados dos FIIs do site Investidor10. '''
     def __init__(self):
         self.base_url = INVESTIDOR10_BASE_URL
@@ -253,7 +224,7 @@ class FIIsScraper:
 
 
 if __name__ == '__main__':
-    FIIsScraper = FIIsScraper()
+    FIIsScraper = Investidor10Scraper()
     fiis = FIIsScraper.get_all_fiis()
     
-    write_csv_file(data=fiis, file_name=FILE_NAME)
+    write_csv_file(data=fiis, file_name='investidor10_fiis.csv')
